+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/character/effects/SwarmWoundsModification.kt
@@ -12,7 +12,7 @@ class SwarmWoundsModification: CharacterEffect {
 
         return character.modifyWounds(
             modifiers.copy(
-                afterMultiplier = modifiers.afterMultiplier * 5.toUInt(),
+                afterMultiplier = modifiers.afterMultiplier * 5,
             )
         )
     }
@@ -26,7 +26,7 @@ class SwarmWoundsModification: CharacterEffect {
 
         return character.modifyWounds(
             modifiers.copy(
-                afterMultiplier = modifiers.afterMultiplier / 5.toUInt(),
+                afterMultiplier = modifiers.afterMultiplier / 5,
             )
         )
     }
